requirements.txt
asyncio_nats_client==0.11.2
pygal==2.4.0
psycopg2_binary==2.8.6
nats==0.1
psycopg2==2.8.6
python-dotenv==0.14.0
